+++ CHANGELOG.rst
@@ -8,6 +8,23 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 .. contents:: ``Release Versions``
 
+`3.9.1`_
+=====================
+
+**Release Date:** ``2025-10-27``
+
+Added
+-----
+
+- Added support to increase the number of DHCP servers from 3 to 16 using the dcnm_network module
+- Added documentation to the dcnm_fabric module for `ENABLE_RT_INTF_STATS` and `INTF_STAT_LOAD_INTERVAL` properties
+
+Fixed
+-----
+
+- https://github.com/CiscoDevNet/ansible-dcnm/issues/518
+
+
 `3.9.0`_
 =====================
 
@@ -643,6 +660,7 @@ The Ansible Cisco Data Center Network Manager (DCNM) collection includes modules
 - cisco.dcnm.dcnm_network	 - Add and remove Networks from a DCNM managed VXLAN fabric.
 - cisco.dcnm.dcnm_interface - DCNM Ansible Module for managing interfaces.
 
+.. _3.9.1: https://github.com/CiscoDevNet/ansible-dcnm/compare/3.9.0...3.9.1
 .. _3.9.0: https://github.com/CiscoDevNet/ansible-dcnm/compare/3.8.1...3.9.0
 .. _3.8.1: https://github.com/CiscoDevNet/ansible-dcnm/compare/3.8.0...3.8.1
 .. _3.8.0: https://github.com/CiscoDevNet/ansible-dcnm/compare/3.7.0...3.8.0